+++ packages/expo-sqlite/CHANGELOG.md
@@ -39,6 +39,10 @@
 
 - Updated SQLite to 3.49.1 and SQLCipher to 4.7.0. ([#35741](https://github.com/expo/expo/pull/35741) by [@kudo](https://github.com/kudo))
 
+### ⚠️ Notices
+
+- Changed async tasks to run on a dedicated parallel queue. ([#35896](https://github.com/expo/expo/pull/35896) by [@kudo](https://github.com/kudo))
+
 ## 15.1.4 - 2025-04-02
 
 ### 🐛 Bug fixes